requirements.txt
pyyaml
pydantic
openai
streamlit
pytest
python-dotenv
